+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## v2.0.2
 * Move file assets from the `assets` to the `ldlib2` folder
+* Fixed cross-OS platform file separator char